+++ src/main/java/me/khajiitos/jackseconomy/screen/WalletScreen.java
@@ -125,7 +125,7 @@ public class WalletScreen extends AbstractContainerScreen<WalletMenu> {
 
     private static void drawCenteredStringNoShadow(GuiGraphics guiGraphics, Font pFont, Component pText, int pX, int pY, int pColor) {
         FormattedCharSequence formattedcharsequence = pText.getVisualOrderText();
-        guiGraphics.drawString(pFont, formattedcharsequence, (int)(pX - pFont.width(formattedcharsequence) / 2), (int)pY, pColor, false);
+        guiGraphics.drawString(pFont, formattedcharsequence, pX - pFont.width(formattedcharsequence) / 2, pY, pColor, false);
     }
 
     @Override
